backend/requirements.txt
Django>=3.2.0,<4.0.0
djangorestframework>=3.12.0,<4.0.0
django-cors-headers>=3.7.0,<4.0.0
djangorestframework-simplejwt>=5.0.0,<6.0.0
psycopg2-binary>=2.9.9